backend/requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
numpy
neo4j
groq
python-dotenv